requirements.txt
bootstrap-flask==2.0.2
click==8.1.2
docx2txt==0.8
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
pdfminer.six==20220319
pdfminer.six[image]
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.27.1
SPARQLWrapper==2.0.0